detect_players_model/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-cloud-vision==3.4.4
opencv-python==4.9.0.80
numpy==1.26.3
requests==2.31.0